feat: Remove rarities with oracle ff code
