+++ .env.example
@@ -7,4 +7,3 @@ PUBLIC_URL=/
 # These are controlled by a remote service, you can still enforce the flags by setting a
 # value here. 1 means enabled, anything else disabled. If the variable is empty it will look the value remotelly.
 REACT_APP_FF_BUILDER_MAINTENANCE=
-REACT_APP_FF_BUILDER_RARITIES_WITH_ORACLE=
