--- src/components/Modals/PublishCollectionModalOld/index.ts
@@ -1,3 +1,0 @@
-import PublishCollectionModal from './PublishCollectionModal.container'
-
-export default PublishCollectionModal
