--- src/components/Modals/PublishCollectionModalWithOracle/index.ts
@@ -1,3 +1,0 @@
-import PublishCollectionModal from './PublishCollectionModal.container'
-
-export default PublishCollectionModal
