+++ src/modules/collection/selectors.ts
@@ -12,7 +12,6 @@ import { CollectionCuration } from 'modules/curations/collectionCuration/types'
 import { getCollectionThirdParty, getData as getThirdParties } from 'modules/thirdParty/selectors'
 import { getThirdPartyForCollection, isUserManagerOfThirdParty } from 'modules/thirdParty/utils'
 import { ThirdParty } from 'modules/thirdParty/types'
-import { getIsRaritiesWithOracleEnabled } from 'modules/features/selectors'
 import { isEqual } from 'lib/address'
 import { isThirdParty } from 'lib/urn'
 import { SET_COLLECTION_MINTERS_SUCCESS, APPROVE_COLLECTION_SUCCESS, REJECT_COLLECTION_SUCCESS } from './actions'
@@ -150,9 +149,6 @@ export const hasViewAndEditRights = (state: RootState, address: string, collecti
  * Returns the corresponding rarities contract depending on if the rarities with oracle
  * feature flag is enabled or not.
  */
-export const getRaritiesContract = (state: RootState, chainId: ChainId) => {
-  const useRaritiesWithOracle = getIsRaritiesWithOracleEnabled(state)
-  const contractName = useRaritiesWithOracle ? ContractName.RaritiesWithOracle : ContractName.Rarities
-
-  return getContract(contractName, chainId)
+export const getRaritiesContract = (chainId: ChainId) => {
+  return getContract(ContractName.RaritiesWithOracle, chainId)
 }
