+++ src/modules/features/types.ts
@@ -1,4 +1,3 @@
 export enum FeatureName {
-  MAINTENANCE = 'maintenance',
-  RARITIES_WITH_ORACLE = 'rarities-with-oracle'
+  MAINTENANCE = 'maintenance'
 }
